generate synth data when we wanna push to redis queue
