+++ validator/control_node/src/main.py
@@ -17,7 +17,7 @@ logger = get_logger(__name__)
 
 async def main() -> None:
     nltk.download('punkt_tab')
-    
+
     config = load_config()
     await config.psql_db.connect()
 
@@ -25,7 +25,7 @@ async def main() -> None:
     await asyncio.gather(
         score_results.main(config),
         sutils.get_save_random_text(),
-        refresh_synthetic_data.main(config),
+        #refresh_synthetic_data.main(config),
         execute_cycle.main(config)
     )
