+++ validator/query_node/src/process_queries.py
@@ -6,8 +6,6 @@ from validator.models import Contender
 from validator.query_node.src.query_config import Config
 from core import task_config as tcfg
 from validator.utils.generic import generic_utils as gutils
-from validator.utils.contender import contender_utils as putils
-from validator.utils.synthetic import synthetic_constants as scst
 from validator.utils.redis import redis_constants as rcst
 from fiber.logging_utils import get_logger
 from validator.utils.redis import redis_dataclasses as rdc
@@ -16,7 +14,6 @@ from validator.db.src.sql.contenders import get_contenders_for_task, update_tota
 from validator.db.src.sql.nodes import get_node
 from validator.utils.generic import generic_constants as gcst
 from opentelemetry import metrics
-import random
 
 logger = get_logger(__name__)
 
@@ -144,8 +141,7 @@ async def process_task(config: Config, message: rdc.QueryQueueMessage):
         logger.debug(f"Successfully acknowledged job id : {message.job_id} ✅")
         await _decrement_requests_remaining(config.redis_db, task)
     else:
-        message.query_payload = await putils.get_synthetic_payload(config.redis_db, task)
-        message.query_payload['seed'] = random.randint(1, scst.MAX_SEED)
+        logger.info("Got synthetic query, processing...")
 
     task_config = tcfg.get_enabled_task_config(task)
     if task_config is None:
@@ -173,13 +169,13 @@ async def process_task(config: Config, message: rdc.QueryQueueMessage):
         raise ValueError("No contenders to query! :(")
 
     COUNTER_TOTAL_QUERIES.add(1, {"task": task, "synthetic_query": str(message.query_type == gcst.SYNTHETIC)})
-    
+
     try:
         if stream:
             return await _handle_stream_query(config, message, contenders_to_query)
         else:
             return await _handle_nonstream_query(config=config, message=message, contenders_to_query=contenders_to_query)
-    
+
     except Exception as e:
         logger.error(f"Error processing task {task}: {e}")
         await _handle_error(
